+++ src/components/core/handlers/downloadHandler.ts
@@ -4,14 +4,15 @@ import {
   PROTOCOL_COMMANDS,
   DownloadURLCommand
 } from '../../../utils/constants.js'
-import { Handler, checkNonce, NonceResponse } from './aHandler.js'
+import { Handler } from './aHandler.js'
+import { checkNonce, NonceResponse } from './utils/nonceHandler.js'
 import { OceanNodeConfig, P2PCommandResponse } from '../../../@types/OceanNode.js'
 import { P2P_CONSOLE_LOGGER } from '../../P2P/index.js'
 import { validateOrderTransaction } from '../validateTransaction.js'
 import { AssetUtils } from '../../../utils/asset.js'
 import { Service } from '../../../@types/DDO/Service.js'
 import { findAndFormatDdo } from './ddoHandler.js'
-import { checkFee } from './feesHandler.js'
+import { checkFee } from './utils/feesHandler.js'
 import { decrypt } from '../../../utils/crypt.js'
 import { Database } from '../../database/index.js'
 import crypto from 'crypto'
